Add util function to AttachmentViewer
