+++ src/shared/services/get-attachment-file-name/index.ts
@@ -1,0 +1,1 @@
+export { getAttachmentFileName } from './getAttachmentFileName'
